core_ca/requirements.txt
flask
pybase64
pyopenssl
requests
gunicorn